pyproject.toml
[tool.poetry]
name = "django-react-typescript-template"
version = "0.1.0"
description = ""
authors = ["Kitty Clock <burzakovskij@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
dj-database-url = "^1.2.0"
django-cors-headers = "^3.13.0"
captcha = "^0.4"
PyJWT = "^2.6.0"
cryptography = "^39.0.0"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
